requirements.txt
fpdf2>=2.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
